requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
plotly==5.17.0
scikit-learn==1.3.0
xgboost==1.7.6
PyMySQL==1.1.0
edge-tts==6.1.9
google-generativeai==0.3.2
psutil==5.9.5
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2